+++ src/BigCommerce/Post_Types/Product/Deletion.php
@@ -5,11 +5,10 @@ namespace BigCommerce\Post_Types\Product;
 
 
 use BigCommerce\Taxonomies\Channel\Channel;
-use BigCommerce\Taxonomies\Channel\Connections;
 
 class Deletion {
 	/**
-	 * @param $post_id
+	 * @param int $post_id
 	 *
 	 * @return void
 	 * @action before_delete_post
@@ -20,15 +19,91 @@ class Deletion {
 		}
 		$product = new Product( $post_id );
 		$bc_id   = $product->bc_id();
-		if ( $bc_id && ! $this->has_posts_in_other_channels( $bc_id, $post_id ) ) {
-			$this->remove_reviews( $bc_id );
+		if ( $bc_id ) {
+			$another_channel_post = $this->get_post_in_another_channel( $bc_id, $post_id );
+			if ( $another_channel_post ) {
+				// re-assign the child images to another post
+				$this->reparent_images( $post_id, $another_channel_post );
+			} else {
+				// the product is complete gone from the site, so get rid of all related imported content
+				$this->remove_reviews( $bc_id );
+				$this->remove_images( $post_id );
+			}
 		}
 	}
 
+	/**
+	 * Remove any reviews related to the product
+	 *
+	 * @param int $product_id
+	 *
+	 * @return void
+	 */
 	private function remove_reviews( $product_id ) {
 		/** @var \wpdb $wpdb */
 		global $wpdb;
 		$wpdb->delete( $wpdb->bc_reviews, [ 'bc_id' => $product_id ], [ '%d' ] );
+	}
+
+	/**
+	 * Remove all images related to the product
+	 *
+	 * @param int $post_id
+	 *
+	 * @return void
+	 */
+	private function remove_images( $post_id ) {
+		$image_ids = $this->identify_child_attachments( $post_id );
+		foreach ( $image_ids as $image ) {
+			wp_delete_attachment( $image, true );
+		}
+	}
+
+	/**
+	 * Set the attachment parent to a new post
+	 *
+	 * @param int $old_parent_id
+	 * @param int $new_parent_id
+	 *
+	 * @return void
+	 */
+	private function reparent_images( $old_parent_id, $new_parent_id ) {
+		$image_ids = $this->identify_child_attachments( $old_parent_id );
+		if ( empty( $image_ids ) || empty( $new_parent_id ) ) {
+			return; // nothing to do
+		}
+		$id_string = implode(',', array_map( 'intval', $image_ids ) );
+
+		/** @var \wpdb $wpdb */
+		global $wpdb;
+		$wpdb->query( $wpdb->prepare( "UPDATE {$wpdb->posts} SET post_parent=%s WHERE post_parent=%s AND ID IN ($id_string)", $new_parent_id, $old_parent_id ) );
+
+		foreach ( $image_ids as $image_id ) {
+			clean_post_cache( $image_id );
+		}
+	}
+
+	/**
+	 * Find all imported attachments with the given post as a parent
+	 *
+	 * @param int $post_id
+	 *
+	 * @return int[]
+	 */
+	private function identify_child_attachments( $post_id ) {
+		return get_posts( [
+			'post_type'      => 'attachment',
+			'post_parent'    => $post_id,
+			'meta_query'     => [
+				[
+					'key'     => 'bigcommerce_id',
+					'compare' => '>',
+					'value'   => 0,
+				],
+			],
+			'fields'         => 'ids',
+			'posts_per_page' => - 1,
+		] );
 	}
 
 	/**
@@ -37,18 +112,12 @@ class Deletion {
 	 * @param int $bc_id
 	 * @param int $post_id
 	 *
-	 * @return bool
+	 * @return int A post ID for the product in another channel
 	 */
-	private function has_posts_in_other_channels( $bc_id, $post_id ) {
-		$connections = new Connections();
-		$channels    = $connections->active();
-		if ( count( $channels ) <= 1 ) {
-			return false;
-		}
-
+	private function get_post_in_another_channel( $bc_id, $post_id ) {
 		$post_channels = get_the_terms( $post_id, Channel::NAME );
 		if ( ! $post_channels ) {
-			return false;
+			return 0;
 		}
 		$post_channel = reset( $post_channels );
 
@@ -73,6 +142,10 @@ class Deletion {
 			],
 		] );
 
-		return ! empty( $matches );
+		if ( empty( $matches ) ) {
+			return 0;
+		}
+
+		return (int) reset( $matches );
 	}
 }