+++ templates/public/components/products/inventory-level.php
@@ -2,17 +2,15 @@
 /**
  * Component: Inventory Level
  *
- * @description Display a message for low stock or out of stock.
+ * @description Display a message for the current inventory level.
  *
  * @var Product $product
+ * @var string  $status The inventory status of the product
+ * @var string  $label  The label for the current stock level
  */
 
 use BigCommerce\Post_Types\Product\Product;
 
-if ( $product->out_of_stock() ) { ?>
-	<span class="bc-product__inventory bc-product__out_of_stock"><?php esc_html_e( '(Out of Stock)', 'bigcommerce' ); ?></span>
-<?php } elseif ( $product->low_inventory() ) {
-	$inventory = $product->get_inventory_level();
-	?>
-	<span class="bc-product__inventory bc-product__low_inventory" ><?php echo esc_html( sprintf( _n( '(%d in Stock)', '(%d in stock)', $inventory, 'bigcommerce' ), $inventory ) ); ?></span>
-<?php } ?>
+?>
+
+<span class="bc-product__inventory bc-product__<?php echo sanitize_html_class( $status ); ?>">(<?php echo esc_html( $label ); ?>)</span>