+++ templates/public/components/products/product-price.php
@@ -5,22 +5,25 @@
  * @description Display the price for a product
  *
  * @var Product $product
+ * @var string  $visible HTML class name to indicate if default pricing should be visible
+ * @var string  $price_range
+ * @var string  $calculated_price_range
  */
 
 use BigCommerce\Post_Types\Product\Product;
 ?>
 <!-- data-js="bc-cached-product-pricing" is required. -->
-<p class="bc-product__pricing--cached bc-product__pricing--visible" data-js="bc-cached-product-pricing">
+<p class="bc-product__pricing--cached <?php echo sanitize_html_class( $visible ); ?>" data-js="bc-cached-product-pricing">
 <?php if ( $product->on_sale() ) { ?>
 	<!-- class="bc-product__original-price" is required. -->
-	<span class="bc-product__original-price"><?php echo esc_html( $product->price_range() ) ?></span>
+	<span class="bc-product__original-price"><?php echo esc_html( $price_range ) ?></span>
 	<!-- class="bc-product__price" is required. -->
 	<span class="bc-product__price bc-product__price--sale">
-		<?php echo esc_html( $product->calculated_price_range() ); ?>
+		<?php echo esc_html( $calculated_price_range ); ?>
 	</span>
 <?php } else { ?>
 	<!-- class="bc-product__price" is required. -->
-	<span class="bc-product__price"><?php echo esc_html( $product->calculated_price_range() ); ?></span>
+	<span class="bc-product__price"><?php echo esc_html( $calculated_price_range ); ?></span>
 <?php } ?>
 </p>